Feature/trello authorization (#53)
* Init authorization

* Add authentication with Trello

* Add field user_id, trelloBoard_id, trello_token to Board model, get boards by token

* Change psycopg2 in requirements_heroku.txt from 2.6.2 to 2.7.1

* Add runtime.txt file to specify Python version

* Update runtime.txt

* Remove runtime.txt

* Update Procfile

* Upgrate version psycopg2==2.7.6

* Fix typo

* Update version django-crispy-forms and django-formtools

* Add readme page

* Show default account

* Fix some small bugs from Author Lujeni revision

+++ core/views/board.py
@@ -13,7 +13,6 @@ from trello import TrelloClient
 from core.models import Board, Webhook
 
 
-trello_client = TrelloClient(api_key=settings.TRELLO_APIKEY, token=settings.TRELLO_TOKEN)
 
 
 @method_decorator(login_required, name='dispatch')
@@ -27,19 +26,27 @@ class BoardView(ListView):
 
     def get(self, request):
         try:
-            boards = trello_client.list_boards()
-            if boards:
-                result = [h.delete() for h in trello_client.list_hooks()]
-                print("delete trello hook :: result={}".format(result))
+            token = request.GET.get('token')
+            user = request.user
+            if token is None:
+                boards = None
+                return super(BoardView, self).get(request)
+            else:
+                trello_client = TrelloClient(api_key=settings.TRELLO_APIKEY, token=token)
+                boards = trello_client.list_boards()
+                if boards:
+                    result = [h.delete() for h in trello_client.list_hooks()]
+                    print("delete trello hook :: result={}".format(result))
 
-            for board in boards:
-                slug_board = slugify(board.name, allow_unicode=False)
-                b, created = Board.objects.get_or_create(name=slug_board)
-                host = getenv("MATTERLLO_HOST") or request.get_host()
-                url = "{}://{}/callback/{}/".format(request.scheme, host, b.id)
-                result = trello_client.create_hook(url, board.id)
-                print("create trello hook :: callback={} :: board={} :: result={}".format(url, slug_board, result))
-            return super(BoardView, self).get(request)
+                for board in boards:
+                    print ("BOARD_ID:", board.id)
+                    slug_board = slugify(board.name, allow_unicode=False)
+                    b, created = Board.objects.get_or_create(name=slug_board, user=user, trello_board_id = board.id, trello_token = token)
+                    host = getenv("MATTERLLO_HOST") or request.get_host()
+                    url = "{}://{}/callback/{}/".format(request.scheme, host, b.id)
+                    result = trello_client.create_hook(url, board.id)
+                    print("create trello hook :: callback={} :: board={} :: result={}".format(url, slug_board, result))
+                return super(BoardView, self).get(request)
         except Exception as e:
             print("unable to display board :: {}".format(e))
             return super(BoardView, self).get(request)
@@ -49,7 +56,14 @@ class BoardView(ListView):
         """
         try:
             context = super(BoardView, self).get_context_data(**kwargs)
+            token = self.request.GET.get('token')
+            if token is None:
+                context['board_list'] = None
+            trello_client = TrelloClient(api_key=settings.TRELLO_APIKEY, token=token)
             trello_client.list_boards()
+            user = self.request.user
+            listboard = Board.objects.filter(trello_token=token)
+            context['board_list'] = listboard
             context['trello_error'] = None
         except Exception as e:
             context['trello_error'] = "{} :: api_key={} :: token={}".format(e,
